+++ docs/apidocs/index.rst
@@ -1,10 +1,10 @@
-.. qiskit_braket_plugin:
+.. qiskit_braket_provider:
 
-.. module:: qiskit_braket_plugin
+.. module:: qiskit_braket_provider
 
-===================================
-Qiskit-Braket Plugin API References
-===================================
+=====================================
+Qiskit-Braket Provider API References
+=====================================
 
 .. toctree::
    :maxdepth: 2
