+++ docs/apidocs/providers.rst
@@ -1,4 +1,4 @@
-.. automodule:: qiskit_braket_plugin.providers
+.. automodule:: qiskit_braket_provider.providers
    :no-members:
    :no-inherited-members:
    :no-special-members:
