+++ docs/index.rst
@@ -1,15 +1,29 @@
-##################################
-Qiskit-Braket plugin documentation
-##################################
+####################################
+Qiskit-Braket provider documentation
+####################################
 
-Qiskit-Braket plugin to execute Qiskit programs on AWS quantum computing hardware devices through Amazon Braket.
+Qiskit-Braket provider to execute Qiskit programs on AWS quantum computing hardware devices through Amazon Braket.
+
+API references
+
+.. toctree::
+  :maxdepth: 2
+
+  API References <apidocs/index>
+
+How-to guides
+
+.. toctree::
+  :maxdepth: 2
+
+  How-to guides <how_tos/index>
+
+Tutorials
 
 .. toctree::
   :maxdepth: 2
 
   Tutorials <tutorials/index>
-  User Guide <how_tos/index>
-  API References <apidocs/index>
 
 .. Hiding - Indices and tables
    :ref:`genindex`
