+++ qiskit_braket_provider/__init__.py
@@ -1,4 +1,4 @@
-"""Qiskit-Braket plugin."""
+"""Qiskit-Braket provider."""
 
 from .providers import (
     AWSBraketProvider,
