+++ tox.ini
@@ -20,7 +20,7 @@ commands =
 envdir = .tox/lint
 commands =
   black --check {posargs} .
-  pylint -rn qiskit_braket_plugin tests
+  pylint -rn qiskit_braket_provider tests
   nbqa pylint docs/ --disable=pointless-statement,missing-module-docstring,invalid-name,expression-not-assigned
   mypy .
 
@@ -33,7 +33,7 @@ basepython = python3
 setenv =
   {[testenv]setenv}
 commands =
-  coverage3 run --source qiskit_braket_plugin --parallel-mode -m pytest
+  coverage3 run --source qiskit_braket_provider --parallel-mode -m pytest
   coverage3 combine
   coverage3 report --fail-under=80
